+++ packages/ui/src/atoms/Input.tsx
@@ -109,6 +109,7 @@ export interface InputProps extends ReactInputProps {
 }
 
 /** Plain input component, can be used in form or outside it */
+// biome-ignore lint/complexity/noExcessiveCognitiveComplexity: <explanation>
 export const Input = React.forwardRef<HTMLInputElement, InputProps>((props, fwRef) => {
   const {
     // default props
@@ -271,7 +272,7 @@ Input.displayName = "Input";
 
 /* Form */
 
-type FormInputProps = Omit<InputProps, "error">;
+type FormInputProps = InputProps;
 
 /** Input field used inside <Form> only. */
 export const FormInput = React.forwardRef<HTMLInputElement, FormInputProps>(